+++ server/src/covet/api/__init__.py
@@ -20,6 +20,7 @@ from covet.api import maintenance as maintenance_router
 from covet.api import manual_bundles as manual_bundles_router
 from covet.api import meta as meta_router
 from covet.api import metadata as metadata_router
+from covet.api import notifications as notifications_router
 from covet.api import photos as photos_router
 from covet.api import share as share_router
 from covet.api import sync as sync_router
@@ -48,6 +49,7 @@ api_router.include_router(metadata_router.router)
 api_router.include_router(documents_router.router)
 api_router.include_router(maintenance_router.router)
 api_router.include_router(chores_router.router)
+api_router.include_router(notifications_router.router)
 api_router.include_router(photos_router.router)
 api_router.include_router(manual_bundles_router.router)
 api_router.include_router(system_router.router)